+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/ConfigurationClientBuilder.java
@@ -72,7 +72,8 @@ import com.azure.data.appconfiguration.models.ConfigurationAudience;
  * <!-- src_embed com.azure.data.applicationconfig.async.configurationclient.instantiation -->
  * <pre>
  * ConfigurationAsyncClient configurationAsyncClient = new ConfigurationClientBuilder&#40;&#41;
- *     .connectionString&#40;connectionString&#41;
+ *     .credential&#40;new DefaultAzureCredentialBuilder&#40;&#41;.build&#40;&#41;&#41;
+ *     .endpoint&#40;endpoint&#41;
  *     .buildAsyncClient&#40;&#41;;
  * </pre>
  * <!-- end com.azure.data.applicationconfig.async.configurationclient.instantiation -->
@@ -82,7 +83,8 @@ import com.azure.data.appconfiguration.models.ConfigurationAudience;
  * <!-- src_embed com.azure.data.applicationconfig.configurationclient.instantiation -->
  * <pre>
  * ConfigurationClient configurationClient = new ConfigurationClientBuilder&#40;&#41;
- *     .connectionString&#40;connectionString&#41;
+ *     .credential&#40;new DefaultAzureCredentialBuilder&#40;&#41;.build&#40;&#41;&#41;
+ *     .endpoint&#40;endpoint&#41;
  *     .buildClient&#40;&#41;;
  * </pre>
  * <!-- end com.azure.data.applicationconfig.configurationclient.instantiation -->
@@ -102,7 +104,7 @@ import com.azure.data.appconfiguration.models.ConfigurationAudience;
  * ConfigurationClient configurationClient = new ConfigurationClientBuilder&#40;&#41;
  *     .pipeline&#40;pipeline&#41;
  *     .endpoint&#40;&quot;https:&#47;&#47;dummy.azure.net&#47;&quot;&#41;
- *     .connectionString&#40;connectionString&#41;
+ *     .credential&#40;new DefaultAzureCredentialBuilder&#40;&#41;.build&#40;&#41;&#41;
  *     .buildClient&#40;&#41;;
  * </pre>
  * <!-- end com.azure.data.applicationconfig.configurationclient.pipeline.instantiation -->